Camera : allow camera to use power key as shutter Provides a way for an app to take control of the power key. Used by the camera to make the power key control the shutter.
Change-Id: I85a1e1761199f4604672be42a3a5005227f5451a
(cherry picked from commit 1566144)

Prevent power key capture when screen is off

The ability for an activity to capture the power key, which was
added to support power key as shutter in the camera, should only
allow the capture when the screen is on. Otherwise, if an activity
that captures the power key is to the front when the device turns
off, the user will be unable to turn it back on.

Change-Id: Ib119d6914ec72554b404c1cc17eef3a932d5d402